Revert "Add a safety check to retroarch-release.sh"
This reverts commit 9ce305940bfb0f0f63d3148e48f3a3c9de0c4a94.

+++ retroarch-release.sh
@@ -5,40 +5,13 @@
 PRGNAM=RetroArch
 SRCNAM="$(printf %s $PRGNAM | tr '[:upper:]' '[:lower:]')"
 TMP=${TMP:-/tmp/libretro}
-FORCE=0
-CLEAN=0
-
-for x in $@; do
-	if [ "$x" == "--force" ]; then
-        FORCE=1
-	fi
-	if [ "$x" == "--clean" ]; then
-        CLEAN=1
-	fi
-done
 
 # Exit on errors and unset variables
 set -eu
 
 # Ensure a clean and fully updated repo
-if [ -d $SRCNAM ]; then
-    if [ $CLEAN -gt 0 ]; then
-		rm -rf -- $SRCNAM
-	elif [ $FORCE -gt 0 ]; then
-		echo "Using existing state of $SRCNAM. If build fails, use --clean to delete and re-clone."
-	else
-		echo "FATAL: $SRCNAM/ exists."
-		echo ""
-		echo " - To build with existing sources: $0 --force"
-		echo " - To delete existing sources and re-clone: $0 --clean"
-		echo ""
-		echo "WARNING: The --clean option does not preserve forks. That is,"
-		echo "the original libretro/$PRGNAM repository will be cloned, not"
-		echo "your personal fork. To build a release build from a fork,"
-		echo "use --force."
-		exit 1
-	fi
-fi
+[ -d $SRCNAM ] && rm -rf -- $SRCNAM
+
 ./libretro-fetch.sh $SRCNAM
 
 COMMIT="$(git --work-tree=$SRCNAM --git-dir=$SRCNAM/.git describe --abbrev=0 \
